backend/pyproject.toml
[project]
name = "noteprinter-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Lukas Winkler",email = "git@lw1.at"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "django-bootstrap5 (>=25.1,<26.0)",
    "django-debug-toolbar (>=5.2.0,<6.0.0)",
    "channels[daphne] (>=4.2.2,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pyvips (>=3.0.0,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-unused-media (>=0.2.2,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
